+++ Common/Orders/LimitOrder.cs
@@ -51,7 +51,7 @@ namespace QuantConnect.Orders
         /// <param name="time">Time the order was placed</param>
         /// <param name="limitPrice">Price the order should be filled at if a limit order</param>
         /// <param name="tag">User defined data tag for this order</param>
-        public LimitOrder(Symbol symbol, int quantity, decimal limitPrice, DateTime time, string tag = "")
+        public LimitOrder(Symbol symbol, decimal quantity, decimal limitPrice, DateTime time, string tag = "")
             : base(symbol, quantity, time, tag)
         {
             LimitPrice = limitPrice;
