+++ Common/Orders/MarketOnOpenOrder.cs
@@ -45,7 +45,7 @@ namespace QuantConnect.Orders
         /// <param name="quantity">The number of units to order</param>
         /// <param name="time">The current time</param>
         /// <param name="tag">A user defined tag for the order</param>
-        public MarketOnOpenOrder(Symbol symbol, int quantity, DateTime time, string tag = "")
+        public MarketOnOpenOrder(Symbol symbol, decimal quantity, DateTime time, string tag = "")
             : base(symbol, quantity, time, tag)
         {
         }
